[frontend] migrate to typescript (#9386)
